Define required features for examples
