+++ examples/http_client_reqwest.rs
@@ -1,0 +1,17 @@
+use opentelemetry::api::trace::Tracer;
+use std::env;
+
+#[tokio::main]
+async fn main() {
+    env_logger::init();
+
+    let instrumentation_key =
+        env::var("INSTRUMENTATION_KEY").expect("env var INSTRUMENTATION_KEY should exist");
+
+    let (tracer, _uninstall) =
+        opentelemetry_application_insights::new_pipeline(instrumentation_key)
+            .with_client(reqwest::Client::new())
+            .install();
+
+    tracer.in_span("reqwest-client", |_cx| {});
+}
