+++ examples/opentelemetry.rs
@@ -72,7 +72,7 @@ async fn main() -> Result<(), Box<dyn Error>> {
         env::var("INSTRUMENTATION_KEY").expect("env var INSTRUMENTATION_KEY should exist");
     let (tracer, _uninstall) =
         opentelemetry_application_insights::new_pipeline(instrumentation_key)
-            .with_client(reqwest::blocking::Client::new())
+            .with_client(reqwest::Client::new())
             .install();
 
     match traceparent {
